backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0
python-multipart==0.0.9
torch==2.3.0
torchvision==0.18.0
Pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
pdf2image==1.17.0
pydicom==2.4.4
reportlab==4.2.0
scipy==1.13.0
pydantic==2.7.3
